Initial test framework for Yalis and precision fix for sampling (#67)
* Support to return logits in generate, logit conversion to float for sampling

* Basic correctness testing

* Increased token generation to 32 and removed dead code

* Fix sample dataset

* Multi GPU tests

+++ yalis/external/sampling.py
@@ -56,7 +56,7 @@ def sample(
         # optionally crop the logits to smallest set of logits with a cumulative probability above top_p
         if top_p < 1.0:
             logits = sample_top_p(logits, top_p)
-        probs = torch.nn.functional.softmax(logits, dim=-1)
+        probs = torch.nn.functional.softmax(logits, dim=-1, dtype=torch.float32)
         return multinomial_num_samples_1(probs)
     return torch.argmax(logits, dim=-1, keepdim=True)
